Structure types to use ws inside wasm
