+++ Cargo.toml
@@ -3,12 +3,19 @@ resolver = "2"
 members = ["crates/*"]
 
 [workspace.dependencies]
+clap = "4"
 chacha20poly1305 = "0.10"
 either = "1.8"
+futures = "0.3"
 rand = { version = "0.8" }
+rmp-serde = {  version = "1" }
+semver = { version = "1.0.14",  features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_bytes = "0.11"
+serde_json = "1"
 serde_with = "2"
+tracing = "0.1"
+tracing-subscriber = "0.3"
 
 [profile.dev.package."*"]
 opt-level = 3
