+++ crates/locutus-core/src/client_events/websocket.rs
@@ -8,11 +8,11 @@ use std::{
 
 use futures::{future::BoxFuture, stream::SplitSink, SinkExt, StreamExt};
 use locutus_runtime::prelude::TryFromTsStd;
+use locutus_stdlib::api::{ClientRequest, ContractRequest, ErrorKind, HostResponse};
 use tokio::sync::mpsc::{channel, Receiver, Sender};
 use warp::{filters::BoxedFilter, Filter, Reply};
 
-use super::{ClientError, ClientEventsProxy, ClientId, ErrorKind, HostResult, OpenRequest};
-use crate::{ClientRequest, ContractRequest, HostResponse};
+use super::{ClientError, ClientEventsProxy, ClientId, HostResult, OpenRequest};
 
 const PARALLELISM: usize = 10; // TODO: get this from config, or whatever optimal way
 
@@ -21,7 +21,7 @@ pub struct WebSocketProxy {
     server_response: Sender<(ClientId, HostResult)>,
 }
 
-type NewResponseSender = Sender<Result<HostResponse<'static>, ClientError>>;
+type NewResponseSender = Sender<Result<HostResponse, ClientError>>;
 
 impl WebSocketProxy {
     /// Starts this as an upgrade to an existing HTTP connection at the `/ws-api` URL
@@ -87,12 +87,11 @@ impl ClientEventsProxy for WebSocketProxy {
         })
     }
 
-    fn send<'a>(
+    fn send(
         &mut self,
         client: ClientId,
-        response: Result<HostResponse<'a>, ClientError>,
+        response: Result<HostResponse, ClientError>,
     ) -> BoxFuture<Result<(), ClientError>> {
-        let response: Result<HostResponse<'static>, _> = response.map(HostResponse::into_owned);
         Box::pin(async move {
             self.server_response
                 .send((client, response))
@@ -262,7 +261,7 @@ async fn new_request(
 
 async fn send_reponse_to_client(
     response_stream: &mut SplitSink<warp::ws::WebSocket, warp::ws::Message>,
-    response: Result<HostResponse<'static>, ClientError>,
+    response: Result<HostResponse, ClientError>,
 ) -> Result<(), Box<dyn std::error::Error>> {
     let serialize = rmp_serde::to_vec(&response).unwrap();
     response_stream
