+++ crates/locutus-dev/Cargo.toml
@@ -11,14 +11,14 @@ repository = "https://github.com/freenet/locutus"
 [dependencies]
 atty = "0.2"
 byteorder = "1"
-clap = { version = "4", features = ["derive", "env"] }
+clap = { workspace = true, features = ["derive", "env"] }
 either = { workspace = true }
-futures = "0.3"
+futures = { workspace = true }
 glob = "0.3"
 thiserror = "1"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3.16", features = ["env-filter", "fmt"] }
-tokio = { version = "1", features = ["rt-multi-thread", "sync", "macros", "fs", "signal", "parking_lot"] }
+tokio = { version = "1", features = ["rt-multi-thread", "sync", "macros", "signal", "parking_lot"] }
 pico-args = "0.5"
 serde = "1"
 serde_json = "1"
@@ -31,11 +31,12 @@ xz2 = "0.1"
 # (mp) message pack feature
 rmp = { version = "0.8", optional = true }
 rmpv = { version = "1", optional = true }
-rmp-serde = { version = "1", optional = true }
+rmp-serde = {  workspace = true }
 
 # internal
 locutus-core = { path = "../locutus-core", version = "0.0.3" }
 locutus-runtime = { path = "../locutus-runtime", version = "0.0.3" }
+locutus-stdlib = { path = "../locutus-stdlib", version = "0.0.3" }
 
 [features]
-messagepack = ["rmp", "rmpv", "rmp-serde"]
+messagepack = ["rmp", "rmpv"]
