+++ crates/locutus-dev/src/local_node/user_events.rs
@@ -7,11 +7,9 @@ use std::{
 
 use either::Either;
 use futures::future::BoxFuture;
-use locutus_core::{
-    ClientError, ClientEventsProxy, ClientId, ClientRequest, ContractRequest, ErrorKind,
-    HostResponse, OpenRequest,
-};
+use locutus_core::{ClientEventsProxy, ClientId, OpenRequest};
 use locutus_runtime::prelude::*;
+use locutus_stdlib::api::{ClientError, ClientRequest, ContractRequest, ErrorKind, HostResponse};
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 
 use crate::{util, CommandSender, DynError};
