+++ crates/locutus-stdlib/Cargo.toml
@@ -16,18 +16,30 @@ byteorder = "1"
 blake2 = { version =  "0.10", features = [ "std" ] }
 bs58 = "0.4"
 env_logger = "0.9"
-log = "0.4"
+futures = { workspace = true }
 once_cell = "1"
 rmpv = { version = "1"}
+rmp-serde = {  workspace = true }
+semver = { workspace = true }
 serde = { workspace = true }
 serde_bytes = { workspace = true }
+serde_json = { workspace = true }
 serde_with = { workspace = true }
 tar = "0.4.38"
 thiserror = "1"
+tracing = {workspace = true }
+tracing-subscriber = { workspace = true, features = ["default", "env-filter"] }
 xz2 = { version = "0.1", optional = true }
 
 # internal
 locutus-macros = { path = "../locutus-macros", version = "0.0.3" }
+
+[target.'cfg(target_family = "unix")'.dependencies]
+tokio = { version = "1", optional = true, features = ["rt-multi-thread", "sync", "macros", "parking_lot"] }
+tokio-tungstenite = { version = "0.17", optional = true } 
+
+[target.'cfg(target_family = "wasm")'.dependencies]
+web-sys = { version = "0.3", features = ["WebSocket"], optional = true }
 
 [dev-dependencies]
 arbitrary = { version = "1", features = ["derive"] }
@@ -38,5 +50,6 @@ rand = { version = "0.8", features = ["small_rng"] }
 
 [features]
 archive = ["xz2"]
+net = ["web-sys", "tokio", "tokio-tungstenite"]
 testing = ["arbitrary"]
 trace = []
