+++ crates/locutus-stdlib/src/contract_interface.rs
@@ -8,15 +8,18 @@ use std::{
     borrow::{Borrow, Cow},
     collections::HashMap,
     fmt::Display,
+    fs::File,
     hash::{Hash, Hasher},
     io::{Cursor, Read},
     ops::{Deref, DerefMut},
+    path::Path,
     str::FromStr,
+    sync::Arc,
 };
 
 use blake2::{Blake2s256, Digest};
 use byteorder::LittleEndian;
-use serde::{Deserialize, Serialize};
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use serde_with::serde_as;
 
 const CONTRACT_KEY_SIZE: usize = 32;
@@ -1176,6 +1179,274 @@ fn internal_fmt_key(
     write!(f, "{}", &r[..8])
 }
 
+/// The state for a contract.
+#[derive(Debug, PartialEq, Eq, Clone, serde::Serialize, serde::Deserialize)]
+#[cfg_attr(feature = "testing", derive(arbitrary::Arbitrary))]
+pub struct WrappedState(
+    #[serde(
+        serialize_with = "WrappedState::ser_state",
+        deserialize_with = "WrappedState::deser_state"
+    )]
+    Arc<Vec<u8>>,
+);
+
+impl WrappedState {
+    pub fn new(bytes: Vec<u8>) -> Self {
+        WrappedState(Arc::new(bytes))
+    }
+
+    pub fn size(&self) -> usize {
+        self.0.len()
+    }
+
+    fn ser_state<S>(data: &Arc<Vec<u8>>, ser: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serde_bytes::serialize(&**data, ser)
+    }
+
+    fn deser_state<'de, D>(deser: D) -> Result<Arc<Vec<u8>>, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let data: Vec<u8> = serde_bytes::deserialize(deser)?;
+        Ok(Arc::new(data))
+    }
+}
+
+impl From<Vec<u8>> for WrappedState {
+    fn from(bytes: Vec<u8>) -> Self {
+        Self::new(bytes)
+    }
+}
+
+impl TryFromTsStd<&rmpv::Value> for WrappedState {
+    fn try_decode(value: &rmpv::Value) -> Result<Self, WsApiError> {
+        let state = value.as_slice().unwrap().to_vec();
+        Ok(WrappedState::from(state))
+    }
+}
+
+impl AsRef<[u8]> for WrappedState {
+    fn as_ref(&self) -> &[u8] {
+        self.0.as_ref()
+    }
+}
+
+impl Deref for WrappedState {
+    type Target = [u8];
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl Borrow<[u8]> for WrappedState {
+    fn borrow(&self) -> &[u8] {
+        &self.0
+    }
+}
+
+impl std::fmt::Display for WrappedState {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let data: String = if self.0.len() > 8 {
+            let last_4 = self.0.len() - 4;
+            self.0[..4]
+                .iter()
+                .map(|b| char::from(*b))
+                .chain("...".chars())
+                .chain(self.0[last_4..].iter().map(|b| char::from(*b)))
+                .collect()
+        } else {
+            self.0.iter().copied().map(char::from).collect()
+        };
+        write!(f, "ContractState(data: [{}])", data)
+    }
+}
+
+/// Just as `locutus_stdlib::Contract` but with some convenience impl.
+#[non_exhaustive]
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct WrappedContract {
+    #[serde(
+        serialize_with = "WrappedContract::ser_contract_data",
+        deserialize_with = "WrappedContract::deser_contract_data"
+    )]
+    pub data: Arc<ContractCode<'static>>,
+    #[serde(
+        serialize_with = "WrappedContract::ser_params",
+        deserialize_with = "WrappedContract::deser_params"
+    )]
+    pub params: Parameters<'static>,
+    pub key: ContractKey,
+}
+
+impl PartialEq for WrappedContract {
+    fn eq(&self, other: &Self) -> bool {
+        self.key == other.key
+    }
+}
+
+impl Eq for WrappedContract {}
+
+impl WrappedContract {
+    pub fn new(data: Arc<ContractCode<'static>>, params: Parameters<'static>) -> WrappedContract {
+        let key = ContractKey::from((&params, &*data));
+        WrappedContract { data, params, key }
+    }
+
+    #[inline]
+    pub fn key(&self) -> &ContractKey {
+        &self.key
+    }
+
+    #[inline]
+    pub fn code(&self) -> &Arc<ContractCode<'static>> {
+        &self.data
+    }
+
+    #[inline]
+    pub fn params(&self) -> &Parameters<'static> {
+        &self.params
+    }
+
+    pub fn get_data_from_fs(path: &Path) -> Result<ContractCode<'static>, std::io::Error> {
+        let mut contract_file = File::open(path)?;
+        let mut contract_data = if let Ok(md) = contract_file.metadata() {
+            Vec::with_capacity(md.len() as usize)
+        } else {
+            Vec::new()
+        };
+        contract_file.read_to_end(&mut contract_data)?;
+        Ok(ContractCode::from(contract_data))
+    }
+
+    fn ser_contract_data<S>(data: &Arc<ContractCode<'_>>, ser: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        data.serialize(ser)
+    }
+
+    fn deser_contract_data<'de, D>(_deser: D) -> Result<Arc<ContractCode<'static>>, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        // let data: ContractCode<'de> = Deserialize::deserialize(deser)?;
+        // Ok(Arc::new(data))
+        todo!()
+    }
+
+    fn ser_params<S>(data: &Parameters<'_>, ser: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        data.serialize(ser)
+    }
+
+    fn deser_params<'de, D>(_deser: D) -> Result<Parameters<'static>, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        // let data: ContractCode<'de> = Deserialize::deserialize(deser)?;
+        // Ok(Arc::new(data))
+        todo!()
+    }
+}
+
+impl<'a> TryFrom<(&'a Path, Parameters<'static>)> for WrappedContract {
+    type Error = std::io::Error;
+    fn try_from(data: (&'a Path, Parameters<'static>)) -> Result<Self, Self::Error> {
+        let (path, params) = data;
+        let data = Arc::new(Self::get_data_from_fs(path)?);
+        Ok(WrappedContract::new(data, params))
+    }
+}
+
+impl TryFromTsStd<&rmpv::Value> for WrappedContract {
+    fn try_decode(value: &rmpv::Value) -> Result<Self, WsApiError> {
+        let contract_map: HashMap<&str, &rmpv::Value> = match value.as_map() {
+            Some(map_value) => HashMap::from_iter(
+                map_value
+                    .iter()
+                    .map(|(key, val)| (key.as_str().unwrap(), val)),
+            ),
+            _ => {
+                return Err(WsApiError::MsgpackDecodeError {
+                    cause: "Failed decoding WrappedContract, input value is not a map".to_string(),
+                })
+            }
+        };
+
+        let key = match contract_map.get("key") {
+            Some(key_value) => ContractKey::try_decode(*key_value).unwrap(),
+            _ => {
+                return Err(WsApiError::MsgpackDecodeError {
+                    cause: "Failed decoding WrappedContract, key not found".to_string(),
+                })
+            }
+        };
+
+        let data = match contract_map.get("data") {
+            Some(contract_data_value) => Arc::new(
+                ContractCode::try_decode(*contract_data_value)
+                    .unwrap()
+                    .into_owned(),
+            ),
+            _ => {
+                return Err(WsApiError::MsgpackDecodeError {
+                    cause: "Failed decoding WrappedContract, data not found".to_string(),
+                })
+            }
+        };
+
+        let params = match contract_map.get("parameters") {
+            Some(params_value) => Parameters::try_decode(*params_value).unwrap().into_owned(),
+            _ => {
+                return Err(WsApiError::MsgpackDecodeError {
+                    cause: "Failed decoding WrappedContract, parameters not found".to_string(),
+                })
+            }
+        };
+
+        Ok(Self { data, params, key })
+    }
+}
+
+impl TryInto<Vec<u8>> for WrappedContract {
+    type Error = ();
+    fn try_into(self) -> Result<Vec<u8>, Self::Error> {
+        Arc::try_unwrap(self.data)
+            .map(|r| r.into_bytes())
+            .map_err(|_| ())
+    }
+}
+
+impl Display for WrappedContract {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "Contract(")?;
+        self.key.fmt(f)?;
+        write!(f, ")")
+    }
+}
+
+#[cfg(feature = "testing")]
+impl<'a> arbitrary::Arbitrary<'a> for WrappedContract {
+    fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
+        use arbitrary::Arbitrary;
+        let data: ContractCode = Arbitrary::arbitrary(u)?;
+        let param_bytes: Vec<u8> = Arbitrary::arbitrary(u)?;
+        let params = Parameters::from(param_bytes);
+        let key = ContractKey::from((&params, &data));
+        Ok(Self {
+            data: Arc::new(data),
+            params,
+            key,
+        })
+    }
+}
+
 #[doc(hidden)]
 pub(crate) mod wasm_interface {
     //! Contains all the types to interface between the host environment and
@@ -1324,12 +1595,12 @@ pub(crate) mod wasm_interface {
         pub fn into_raw(self) -> i64 {
             #[cfg(feature = "trace")]
             {
-                log::trace!("returning FFI -> {self:?}");
+                tracing::trace!("returning FFI -> {self:?}");
             }
             let ptr = Box::into_raw(Box::new(self));
             #[cfg(feature = "trace")]
             {
-                log::trace!("FFI result ptr: {ptr:p} ({}i64)", ptr as i64);
+                tracing::trace!("FFI result ptr: {ptr:p} ({}i64)", ptr as i64);
             }
             ptr as _
         }
@@ -1341,7 +1612,7 @@ pub(crate) mod wasm_interface {
             )));
             #[cfg(feature = "trace")]
             {
-                log::trace!(
+                tracing::trace!(
                     "got FFI result @ {ptr} ({:p}) -> {result:?}",
                     ptr as *mut Self
                 );
@@ -1351,7 +1622,7 @@ pub(crate) mod wasm_interface {
 
         #[cfg(feature = "trace")]
         fn log_input<T: std::fmt::Debug>(&self, serialized: &[u8], value: &T, ptr: *mut u8) {
-            log::trace!(
+            tracing::trace!(
                 "got result through FFI; addr: {:p} ({}i64, mapped: {ptr:p})
                  serialized: {serialized:?}
                  value: {value:?}",
@@ -1373,7 +1644,7 @@ pub(crate) mod wasm_interface {
                     let size = serialized.len() as _;
                     let ptr = serialized.as_ptr();
                     #[cfg(feature = "trace")] {
-                        log::trace!(
+                        tracing::trace!(
                             "sending result through FFI; addr: {ptr:p} ({}),\n  serialized: {serialized:?}\n  value: {value:?}",
                             ptr as i64
                         );
