+++ crates/locutus-stdlib/src/lib.rs
@@ -1,7 +1,10 @@
 //! Standard library provided by the Freenet project to be able to write Locutus-compatible contracts.
+#[cfg(feature = "net")]
+pub mod api;
 pub mod buf;
 mod component_interface;
 pub mod contract_interface;
+mod versioning;
 #[cfg(feature = "xz2")]
 pub mod web;
 
@@ -12,12 +15,14 @@ pub mod prelude {
     pub use crate::component_interface::*;
     pub use crate::contract_interface::wasm_interface::*;
     pub use crate::contract_interface::*;
+    pub use crate::versioning::*;
     pub use locutus_macros::{component, contract};
 
     pub use bincode;
     pub use blake2;
     pub use env_logger;
-    pub use log;
+    pub use tracing;
+    pub use tracing_subscriber;
 }
 
 #[doc(hidden)]
